requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
google-cloud-firestore>=2.9
pytest>=7.0
requests>=2.28
python-dotenv>=1.0
